Continue runtime separation
