+++ runtime/Cargo.toml
@@ -326,7 +326,6 @@ try-runtime = [
     "pallet-author-slot-filter?/try-runtime",
     "parachain-staking?/try-runtime",
 ]
-without-sudo=[]
 
 [package]
 authors = ["Zeitgeist PM <contact@zeitgeist.pm>"]
